Update topics.models.js ::INTEGER on comment-count

+++ models/topics.models.js
@@ -18,7 +18,7 @@ exports.selectArticles = (topic, sort_by, order) => {
   articles.votes,
   comment_count,
   article_img_url, 
-  COUNT(comments.comment_id) AS comment_count
+  COUNT(comments.comment_id)::INTEGER AS comment_count
   FROM articles
   JOIN comments ON articles.article_id = comments.article_id`;
